Use 1Hz refresh timers with absolute timestamps
The application's event loop has been "ticking" at 10Hz,
determining the rate at which the seconds count in the main clock
is checked for accuracy. This commit changes the "tick" function to
instead set timers with urwid's set_alarm_at() function with
absolute time stamps, scheduling the clock display update at a rate
of 1Hz with updates timed to the top of each second.